feat: add customer handlers, repository e ajustes service user update

+++ internal/core/port/customer.go
@@ -10,7 +10,7 @@ type CustomerRepository interface {
 	CreateCustomer(ctx context.Context, customer *domain.Customer) (*domain.Customer, error)
 	GetCustomerByID(ctx context.Context, id string) (*domain.Customer, error)
 	GetCustomerByName(ctx context.Context, name string) (*domain.Customer, error)
-	ListCustomers(ctx context.Context, page, limit int64) ([]domain.Customer, error)
+	ListCustomers(ctx context.Context, page, limit int) ([]domain.Customer, error)
 	UpdateCustomer(ctx context.Context, customer *domain.Customer) (*domain.Customer, error)
 	DeleteCustomer(ctx context.Context, id string) error
 }
@@ -18,7 +18,7 @@ type CustomerRepository interface {
 type CustomerService interface {
 	CreateCustomer(ctx context.Context, customer *domain.Customer) (*domain.Customer, error)
 	GetCustomer(ctx context.Context, id string) (*domain.Customer, error)
-	ListCustomers(ctx context.Context, page, limit int64) ([]domain.Customer, error)
+	ListCustomers(ctx context.Context, page, limit int) ([]domain.Customer, error)
 	UpdateCustomer(ctx context.Context, customer *domain.Customer) (*domain.Customer, error)
 	DeleteCustomer(ctx context.Context, id string) error
 }
